mmc: jz4740: rename to jz47xx_mmc
The MMC controllers in later Ingenic JZ47xx SoCs (e.g. the JZ4780) are
similar to the one in the 4740, with some additional functionality.
Therefore, the same driver can be used for all of them, with a few
minor alterations.

This patch renames the existing jz4740_mmc driver to jz47xx_mmc, in
preparation for the addition of support for the JZ4780. The platform
data struct definition has been moved from asm/mach-jz4740/jz4740_mmc.h
to linux/platform_data/mmc-jz47xx.h. An id_table has been added to the
platform_driver which will match against the original "jz4740-mmc" name
(and later "jz4780-mmc").

There still remains a bit of JZ4740-specific code to configure the GPIO
pins, conditionally compiled on CONFIG_MACH_JZ4740. Ideally this should
be moved to platform code in future.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>