mmc: jz47xx: add DMA support
Add support for using the DMA engine for MMC data transfers. PIO is used
as a fallback if no DMA channel can be obtained, or if setting up a DMA
transfer fails.

Currently, the DMA channel is requested using dma_request_slave_channel().
This means that DMA will only be used on platforms which use a device tree
and specify a DMA channel for the MMC controller, i.e. jz4780. However,
the code has been written with older SoCs in mind and should work on the
4740 if somebody with the appropriate hardware were to implement DMA
channel request for non-DT platforms and test it.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>