drm: gem: use dma_alloc_coherent instead of write_combine
Not intended for upstream.
No need for review.

This is a hack for 3.14/15

There is talk on the lists about moving dma_alloc_write_combine
to a generic api and add wrappers.

mips doesn't have that at the moment. This simple fix allows us
to proceed with the drm hdmi driver without any big hassle.