gpu: drm: jz4780: Add LCD and HDMI driver
Adds LCD and HDMI driver for the jz4780.

The jz4780 has the same hdmi block as the imx6 processor.
The hdmi driver is based on that from staging. Howver it has been
modified to make it more generic and usable by other SoCs.

A phandle from the hdmi node and i2c node is needed by the jz4780
lcd driver to initialize the encoder/connector. i2c node is required
for probe deferral only.

Note: The jz4780 LCDC likes the framebuffer to be explicitly unpacked.
This is especially necessary for the dma to work.

Hence the DefaultFbBpp must be set to 32. Otherwise XOrg defaults to
24 and the driver doesn't work.

Sample Xorg.conf for modesetting

Section "Device"
        Identifier      "Video Device"
        Driver          "modesetting"
EndSection

Section "Monitor"
        Identifier      "Monitor"
        Option          "PreferredMode" "1920x1080"
EndSection

Section "Screen"
        Identifier      "Screen"
        Monitor         "Monitor"
        Device          "Video Device"
        DefaultFbBpp    32
EndSection

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>