mmc: jz47xx: wait for dma transaction to complete before checking done
Currently, the driver sets up and starts the dma transaction,
and then immediately starts polling the done state in the mmc core.

This patch adds a completion callback to wait for the dma transaction
to complete before checking the done state.

Wifi on the CI20 uses the brcmfmac-sdio driver which uses this driver.

During pings (by extension all Tx/Rx), part of the data packet would
return back to userland correctly. With this fix, the complete data is
read back from the wifi chip.

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>