rtc: jz47xx: change to core_initcall for iw8103
This patch changes the rtc driver init sequence to register
using core_initcall. This is done in preparation for adding
the IW8103 driver.

The IW8103 driver requires that the rtc clock provider be available
before it registers.

Simple EPROBE_DEFER in IW8103 doesn't work as the platform_device_register
call seems to miss registering the subsequent brmfmac-sdio driver.

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>