dma: jz4780: fix bug in jz4780_dma_slave_config
Currently this function of the dma driver has two bugs.

1) src_addr_width is checked twice for DMA_SLAVE_BUSWIDTH_UNDEFINED and
dst_addr_width is not checked for DMA_SLAVE_BUSWIDTH_UNDEFINED at all.

2) If there is a dma transaction which is unidirectional. This function
doesn't allow it. Because at the moment. If any src/dst addr width is
8 byte or zero (undefined), then the function returns -einval.

In practice, e.g. pcm audio driver for playback. Sets up the dma for a
unidirectional transfer. -EINVAL is returned. Which should not be the case.

This patch fixes them both.

First checking if either src or dst addr width is 8bytes, then return -EINVAL.

And then check if both src and dst addrwidth are undefined, then return -EINVAL.

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>
Reviewed-by: Alex Smith <alex@alex-smith.me.uk>