dt: ci20: Add wifi regulator and msc1 node to dt
The wifi chip on the CI20 is the IW8103 which is a rebranded brcm4330.
The chip uses the brmfmac-sdio driver.

This patch adds 2 fixed regulators to control the power and reset line.

Reset is always held high to keep the chip active.
Power regulator is controlled by the mmc core using vmmc-supply.

The IW8103 is automatically detected by the mmc driver and the brcmfmac-sdio
driver registers the chip.

The driver has been tested with firmware and nvram loaded into the kernel.

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>