dt: ci20: Add wifi regulator and msc1 node to dt
The wifi chip on the CI20 is the IW8103 which is a rebranded brcm4330.
The chip uses the brmfmac-sdio driver.

If power and reset are correctly set, the IW8103 is automatically
detected by the mmc driver as a card and the brcmfmac-sdio driver
registers the chip.

This patch adds 2 fixed regulators to control the power and reset line.

Power regulator is controlled by the mmc core using vmmc-supply.

Reset is always held high to keep the chip active. This is a hack
as there is no easy way to do this using the brcmfmac-sdio driver.
Or the mmc driver that it relies on.

The driver has been tested with firmware and nvram loaded into the kernel.

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>