mmc: jz47xx: Add vmmc-supply support
The current mmc driver defines its own ingenic,power-gpio for power
handling. Although it doesn't use the gpio much in the driver and it
is an optional binding.

The mmc core supports power regulator using the vmmc-supply binding.

This patch adds vmmc-supply support to the driver and removes
the ingenic,power-gpio

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>