sound: soc: codec: Add codec driver for Ingenic JZ4780
This patch adds a codec driver for the Ingenic JZ4780 SoC
The internal codec headphone output is supported.

HDMI output is not supported in the driver yet. This will have to
wait until the HDMI driver supports audio.

Input through a microphone connected as part of a 4-pin OMTP-compliant
headset is also supported

Tested with various combinations of VLC, Gnome sound recorder & mplayer

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>