sound: soc: jz4780: Add audio driver for the CI20 board
This patch adds an ALSA SoC audio driver for the CI20 & the Ingenic
JZ4780 SoC that it's built around.

The internal codec headphone output is supported.
HDMI output is not supported in the driver until dwc-hdmi driver
supports audio.

Input through a microphone connected as part of a 4-pin OMTP-compliant
headset is also supported

Tested with various combinations of VLC, Gnome sound recorder & mplayer

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>