MIPS: mm: c-r4k: Ensure CCA is set to non-coherent on UP kernels.
In case the YAMON bootloader is being used, it will set the Config0/K0
value to 0x5 if a multicore processor is detected. This may have
undesired effects if a CM is present since certain cache operations
may broadcast and go through the CM even if we run a UP kernel.
Therefore it's best to ensure that we are in non coherent mode
in UP kernels even if CM is present.

Signed-off-by: Markos Chandras <markos.chandras@imgtec.com>