irq: jz4780: do not pass jz4780_intc structure pointer as handler data
The irq handler data is used to pass the pointer to the irq structure.
Do not use that and use a variable instead.

This is necessary to use the jz4740-adc driver which uses the handler
data to pass data to the chained irq handler of a bunch of virtual
irq muxed behind this hw irq.

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>