ci20: HACK: Use fixed regulators to keep BT always ON.
The bluetooth on the CI20 is on the same chip as the wifi.
The interface is UART. But the power, reset, wake I/O
need to be configured for the chip to function.

There is no driver as such.
After these I/Os are set, the broadcom patchram utility works
and loads the firmware on the chip.

The bluetooth service attaches to the chip and the device starts
scanning.